move `loss` and `metrics` to `nn`
