+++ chemprop/v2/nn/metrics.py
@@ -5,8 +5,8 @@ import torch
 from torch import Tensor
 from torchmetrics import functional as F
 
-from chemprop.v2.utils.registry import ClassRegistry
-from chemprop.v2.models.loss import BCELoss, CrossEntropyLoss, LossFunction, MSELoss
+from chemprop.v2.utils import ClassRegistry
+from chemprop.v2.nn.loss import BCELoss, CrossEntropyLoss, LossFunction, MSELoss
 
 MetricRegistry = ClassRegistry()
